Во всех контрактах заменил decimal на double, т.к. монга не умеет нормально сериализовать decimal

+++ TeachMe/Controllers/AccountController.cs
@@ -167,7 +167,7 @@ namespace TeachMe.Controllers
             return View(model);
         }
 
-        private decimal GetInitialUserCash()
+        private double GetInitialUserCash()
         {
             return ProjectType == ProjectType.Student ? ApplicationSettings.StudentInitialCash : ApplicationSettings.TeacherInitialCash;
         }
